feat: optimize jumping strategy with jump2d
Here is my reasoning for the settings:

- I don't like the defaults that come with jump2d as it creates far too
  much visual noise on the screen, so I use the single character option.
  This also allows me to jump to _any_ location vs only the start/end of
  words in the default configuration.

- With single character mode, I've expanded the labels to include caps
  as well. Typing a capital letter is still faster than typing two lower
  case letters as its a chord versus two sequential presses. With this
  change alone, I'm able to get to most places with 3 presses: trigger
  key, character of interest, and one label.

- In the cases where the above is not sufficiently unique and requires
  an additional label, I want that label showed to me in advance to
  reduce the mental delay.

With folke releasing flash 2.0, I wanted to give my previous jumper
another look, but in the end, while it's an amazing plug-in, it's also
five times the amount of code. The one feature I do miss are the
treesitter selection jumps, but I wonder if I could write my own spotter
function for jump2d.