feat: replace mini.pairs with autopairs
I don't understand why people continue to use mini.pairs as it is
inferior to nvim-autopairs in my experience. With mini.pairs, I feel
like I'm constantly fighting with it because it adds closing quotes
where it should not, it doesn't handle triple quotes (markdown code
blocks or python docstrings), which are a very large part of my life.