fix: move monkey patch of picker.preview.file to util.lua
We cannot monkey patch picker.preview.file from the snacks.picker.config
function anymore as that function is called multiple times, so we just
end up wrapping the wrapper over and over again.

We also have to change the buftype to `nofile` as folke changed that
behavior in subsequent versions where buftype is ``. Not sure why he did
that, but this was causing the LSP to fire up, which I do not want.