fix(minifiles): use hl namespace to highlight target window
The previous implementation used 'winhighlight' option to customize the
hl groups in the preview window. That option is a window local option,
which has the following consequence (from docs):

“The options local to a window are remembered for each buffer.”

As a consequence, when a user opened a file in the explorer, our restore
function to remove the window local option was applied to the correct
window, but it was already hosting the new file. So when one would flip
back to the prior buffer, the window local options were still applied.